ibm-qiskit/requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
cvxopt==1.2.5.post1
decorator==4.4.2
dlx==1.0.4
docplex==2.15.194
fastdtw==0.3.4
idna==2.10
inflection==0.5.1
joblib==0.17.0
jsonschema==2.6.0
marshmallow==2.21.0
marshmallow-polyfield==3.2
more-itertools==8.5.0
mpmath==1.1.0
nest-asyncio==1.4.1
networkx==2.5
ntlm-auth==1.5.0
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
ply==3.11
psutil==5.7.2
pycparser==2.20
pylatexenc==2.7
python-dateutil==2.8.1
pytz==2020.1
qiskit-aer==0.2.3
qiskit-aqua==0.5.2
qiskit-ibmq-provider==0.3.0
qiskit-ignis==0.1.1
qiskit-terra==0.8.2
Quandl==3.5.3
requests==2.24.0
requests-ntlm==1.1.0
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
sympy==1.6.2
threadpoolctl==2.1.0
urllib3==1.25.11
websockets==7.0
